test(server): rename and extract one test from MtvRoomPositionConstraintAndDevicePositionUpdate
